test(cpu-x86): a guest that enters long mode and runs 64-bit code
Twenty-three directed tests, written from the *Intel SDM* volume 3 §9.8.5
and volume 2's encoding rules and from the *AMD64 Architecture
Programmer's Manual* volume 2 where AMD is clearer. Each cites its
volume and section where the behaviour is not obvious.

The one that matters is `a_guest_enters_long_mode_and_executes_64_bit_code`:
nothing is set up from outside except the page tables in memory. The
processor starts in 32-bit protected mode and executes the activation
sequence itself — `CR4.PAE`, `CR3`, `EFER.LME` through `WRMSR`, `CR0.PG`,
then a far jump to a code segment with `L` set — and then runs a 64-bit
immediate load, a `REX.B` move into `R15`, a sign-extended `imm32` and a
`RIP`-relative store.

The rest take that apart: the ordering rules the activation sequence
enforces, `LMA` being the processor's bit and not software's, the
zero-extend/preserve asymmetry, `REX` renaming the high-byte registers,
`RIP`-relative counting from the end of the instruction, the `d64` stack
and branch widths, `MOVSXD`, the twenty encodings long mode reclaimed,
the canonical check, sixteen-byte gates with `IRETQ`, `SYSCALL`/`SYSRET`,
`SWAPGS`, no-execute, 1 GiB pages, PAE without long mode, and a snapshot
round trip that would otherwise lose `EFER`.

+++ src/cpu/x86/exec.rs
@@ -721,9 +721,9 @@ impl<'a> Exec<'a> {
     fn legacy_read16_seg(&mut self, segment: u16, offset: u16) -> u16 {
         let base = linear(segment, offset);
         if self.word_is_one_cycle(base, offset) {
-            return self.phys_read(u64::from(base), 2) as u16;
+            return self.phys_read(base, 2) as u16;
         }
-        let lo = self.phys_read(u64::from(base), 1) as u8;
+        let lo = self.phys_read(base, 1) as u8;
         let hi = self.phys_read(linear(segment, offset.wrapping_add(1)), 1) as u8;
         u16::from(lo) | (u16::from(hi) << 8)
     }
@@ -732,10 +732,10 @@ impl<'a> Exec<'a> {
     fn legacy_write16_seg(&mut self, segment: u16, offset: u16, value: u16) {
         let base = linear(segment, offset);
         if self.word_is_one_cycle(base, offset) {
-            self.phys_write(u64::from(base), 2, u64::from(value));
+            self.phys_write(base, 2, u64::from(value));
             return;
         }
-        self.phys_write(u64::from(base), 1, u64::from(value & 0xff));
+        self.phys_write(base, 1, u64::from(value & 0xff));
         self.phys_write(
             linear(segment, offset.wrapping_add(1)),
             1,
